move Context to separate file
