+++ parser/src/json/schema.rs
@@ -1,20 +1,21 @@
-use crate::{HashMap, HashSet};
+use crate::HashMap;
 use anyhow::{anyhow, bail, Result};
 use derivre::RegexAst;
 use indexmap::{IndexMap, IndexSet};
-use referencing::{Draft, Registry, Resolver, Resource, ResourceRef, Retrieve};
+use referencing::{Draft, ResourceRef, Retrieve};
 use serde_json::Value;
-use std::{any::type_name_of_val, cell::RefCell, mem, rc::Rc};
+use std::mem;
 
+use super::context::{Context, PreContext};
 use super::formats::lookup_format;
 use super::numeric::Decimal;
 
-const DEFAULT_ROOT_URI: &str = "json-schema:///";
-const DEFAULT_DRAFT: Draft = Draft::Draft202012;
+pub use super::context::RetrieveWrapper;
+
 const TYPES: [&str; 6] = ["null", "boolean", "number", "string", "array", "object"];
 
 // Keywords that are implemented in this module
-const IMPLEMENTED: [&str; 24] = [
+pub const IMPLEMENTED: [&str; 24] = [
     // Core
     "anyOf",
     "oneOf",
@@ -49,7 +50,7 @@ const IMPLEMENTED: [&str; 24] = [
 // Keywords that are used for metadata or annotations, not directly driving validation.
 // Note that some keywords like $id and $schema affect the behavior of other keywords, but
 // they can safely be ignored if other keywords aren't present
-const META_AND_ANNOTATIONS: [&str; 15] = [
+pub const META_AND_ANNOTATIONS: [&str; 15] = [
     "$anchor",
     "$defs",
     "definitions",
@@ -447,130 +448,14 @@ impl Schema {
 }
 
 #[derive(Clone)]
-struct SchemaBuilderOptions {
-    max_size: usize,
+pub struct SchemaBuilderOptions {
+    pub max_size: usize,
 }
 
 impl Default for SchemaBuilderOptions {
     fn default() -> Self {
         SchemaBuilderOptions { max_size: 50_000 }
     }
-}
-
-struct SharedContext {
-    defs: HashMap<String, Schema>,
-    seen: HashSet<String>,
-    n_compiled: usize,
-}
-
-impl SharedContext {
-    fn new() -> Self {
-        SharedContext {
-            defs: HashMap::default(),
-            seen: HashSet::default(),
-            n_compiled: 0,
-        }
-    }
-}
-
-struct Context<'a> {
-    resolver: Resolver<'a>,
-    draft: Draft,
-    shared: Rc<RefCell<SharedContext>>,
-    options: SchemaBuilderOptions,
-}
-
-impl<'a> Context<'a> {
-    fn in_subresource(&'a self, resource: ResourceRef) -> Result<Context<'a>> {
-        let resolver = self.resolver.in_subresource(resource)?;
-        Ok(Context {
-            resolver: resolver,
-            draft: resource.draft(),
-            shared: Rc::clone(&self.shared),
-            options: self.options.clone(),
-        })
-    }
-
-    fn as_resource_ref<'r>(&'a self, contents: &'r Value) -> ResourceRef<'r> {
-        self.draft
-            .detect(contents)
-            .unwrap_or(DEFAULT_DRAFT)
-            .create_resource_ref(contents)
-    }
-
-    fn normalize_ref(&self, reference: &str) -> Result<String> {
-        Ok(self
-            .resolver
-            .resolve_against(&self.resolver.base_uri().borrow(), reference)?
-            .normalize()
-            .into_string())
-    }
-
-    fn lookup_resource(&'a self, reference: &str) -> Result<ResourceRef<'a>> {
-        let resolved = self.resolver.lookup(reference)?;
-        Ok(self.as_resource_ref(&resolved.contents()))
-    }
-
-    fn insert_ref(&self, uri: &str, schema: Schema) {
-        self.shared
-            .borrow_mut()
-            .defs
-            .insert(uri.to_string(), schema);
-    }
-
-    fn get_ref_cloned(&self, uri: &str) -> Option<Schema> {
-        self.shared.borrow().defs.get(uri).cloned()
-    }
-
-    fn mark_seen(&self, uri: &str) {
-        self.shared.borrow_mut().seen.insert(uri.to_string());
-    }
-
-    fn been_seen(&self, uri: &str) -> bool {
-        self.shared.borrow().seen.contains(uri)
-    }
-
-    fn is_valid_keyword(&self, keyword: &str) -> bool {
-        if !self.draft.is_known_keyword(keyword)
-            || IMPLEMENTED.contains(&keyword)
-            || META_AND_ANNOTATIONS.contains(&keyword)
-        {
-            return true;
-        }
-        return false;
-    }
-
-    fn increment(&self) -> Result<()> {
-        let mut shared = self.shared.borrow_mut();
-        shared.n_compiled += 1;
-        if shared.n_compiled > self.options.max_size {
-            bail!("schema too large");
-        }
-        Ok(())
-    }
-}
-
-#[derive(Clone)]
-pub struct RetrieveWrapper(pub Rc<dyn Retrieve>);
-impl RetrieveWrapper {
-    pub fn new(retrieve: Rc<dyn Retrieve>) -> Self {
-        RetrieveWrapper(retrieve)
-    }
-}
-impl std::ops::Deref for RetrieveWrapper {
-    type Target = dyn Retrieve;
-    fn deref(&self) -> &Self::Target {
-        self.0.as_ref()
-    }
-}
-impl std::fmt::Debug for RetrieveWrapper {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", type_name_of_val(&self.0))
-    }
-}
-
-fn draft_for(value: &Value) -> Draft {
-    DEFAULT_DRAFT.detect(value).unwrap_or(DEFAULT_DRAFT)
 }
 
 pub fn build_schema(
@@ -585,34 +470,12 @@ pub fn build_schema(
         }
     }
 
-    let draft = draft_for(&contents);
-    let resource = draft.create_resource(contents);
-    let base_uri = resource.id().unwrap_or(DEFAULT_ROOT_URI).to_string();
+    let pre_ctx = PreContext::new(contents, retriever)?;
+    let ctx = Context::new(&pre_ctx)?;
 
-    let registry = {
-        // Weirdly no apparent way to instantiate a new registry with a retriever, so we need to
-        // make an empty one and then add the retriever + resource that may depend on said retriever
-        let empty_registry =
-            Registry::try_from_resources(std::iter::empty::<(String, Resource)>())?;
-        empty_registry.try_with_resources_and_retriever(
-            vec![(&base_uri, resource)],
-            retriever.unwrap_or(&referencing::DefaultRetriever),
-            draft,
-        )?
-    };
-
-    let resolver = registry.try_resolver(&base_uri)?;
-    let ctx = Context {
-        resolver: resolver,
-        draft: draft,
-        shared: Rc::new(RefCell::new(SharedContext::new())),
-        options: SchemaBuilderOptions::default(),
-    };
-
-    let root_resource = ctx.lookup_resource(&base_uri)?;
+    let root_resource = ctx.lookup_resource(&pre_ctx.base_uri)?;
     let schema = compile_resource(&ctx, root_resource)?;
-    let defs = std::mem::take(&mut ctx.shared.borrow_mut().defs);
-    Ok((schema, defs))
+    Ok((schema, ctx.take_defs()))
 }
 
 fn compile_resource(ctx: &Context, resource: ResourceRef) -> Result<Schema> {
